refactor(python): move compatible converters
