+++ python/pyfory/struct.py
@@ -500,9 +500,12 @@ class DataClassSerializer(Serializer):
             else {}
         )
         self._missing_field_defaults = self._build_missing_field_defaults()
+        from pyfory.converter import CompatibleScalarFieldSerializer
+
+        self._compatible_scalar_flags = [isinstance(serializer, CompatibleScalarFieldSerializer) for serializer in self._serializers]
         self._basic_field_flags = [
             (not self._dynamic_fields.get(field_name, False))
-            and not getattr(self._serializers[index], "_compatible_scalar_conversion", False)
+            and not self._compatible_scalar_flags[index]
             and isinstance(self._serializers[index], self._BASIC_SERIALIZERS)
             for index, field_name in enumerate(self._field_names)
         ]
@@ -554,7 +557,7 @@ class DataClassSerializer(Serializer):
         else:
             write_context.write_no_ref(field_value, serializer=serializer)
 
-    def _read_field_value(self, read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref):
+    def _read_field_value(self, read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref, is_compatible_scalar):
         if is_nullable and is_basic:
             if read_context.read_int8() == NULL_FLAG:
                 return None
@@ -567,7 +570,7 @@ class DataClassSerializer(Serializer):
             flag = read_context.read_int8()
             if flag == NULL_FLAG:
                 return None
-            if getattr(serializer, "_compatible_scalar_conversion", False) and flag != NOT_NULL_VALUE_FLAG:
+            if is_compatible_scalar and flag != NOT_NULL_VALUE_FLAG:
                 from pyfory.error import ForyInvalidDataError
 
                 raise ForyInvalidDataError(f"Invalid compatible scalar null flag: {flag}")
@@ -662,10 +665,27 @@ class DataClassSerializer(Serializer):
                 is_dynamic = self._dynamic_fields.get(field_name, False)
                 is_tracking_ref = self._ref_fields.get(field_name, False)
                 is_basic = self._basic_field_flags[index]
+                is_compatible_scalar = self._compatible_scalar_flags[index]
                 if field_name not in self._current_class_field_names or not self._assign_fields[index]:
-                    self._read_missing_field_value(read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref)
+                    self._read_missing_field_value(
+                        read_context,
+                        serializer,
+                        is_nullable,
+                        is_dynamic,
+                        is_basic,
+                        is_tracking_ref,
+                        is_compatible_scalar,
+                    )
                     continue
-                field_value = self._read_field_value(read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref)
+                field_value = self._read_field_value(
+                    read_context,
+                    serializer,
+                    is_nullable,
+                    is_dynamic,
+                    is_basic,
+                    is_tracking_ref,
+                    is_compatible_scalar,
+                )
                 validation_field_type = self._validation_field_types[index]
                 if validation_field_type is None:
                     interned_name = self._field_name_interned[field_name]
@@ -683,7 +703,16 @@ class DataClassSerializer(Serializer):
                     is_dynamic = self._dynamic_fields.get(field_name, False)
                     is_tracking_ref = self._ref_fields.get(field_name, False)
                     is_basic = self._basic_field_flags[index]
-                    field_value = self._read_field_value(read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref)
+                    is_compatible_scalar = self._compatible_scalar_flags[index]
+                    field_value = self._read_field_value(
+                        read_context,
+                        serializer,
+                        is_nullable,
+                        is_dynamic,
+                        is_basic,
+                        is_tracking_ref,
+                        is_compatible_scalar,
+                    )
                     interned_name = self._field_name_interned[field_name]
                     if obj_dict is not None:
                         obj_dict[interned_name] = field_value
@@ -696,7 +725,16 @@ class DataClassSerializer(Serializer):
                     is_dynamic = self._dynamic_fields.get(field_name, False)
                     is_tracking_ref = self._ref_fields.get(field_name, False)
                     is_basic = self._basic_field_flags[index]
-                    field_value = self._read_field_value(read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref)
+                    is_compatible_scalar = self._compatible_scalar_flags[index]
+                    field_value = self._read_field_value(
+                        read_context,
+                        serializer,
+                        is_nullable,
+                        is_dynamic,
+                        is_basic,
+                        is_tracking_ref,
+                        is_compatible_scalar,
+                    )
                     validation_field_type = self._validation_field_types[index]
                     if validation_field_type is None:
                         interned_name = self._field_name_interned[field_name]
@@ -717,11 +755,28 @@ class DataClassSerializer(Serializer):
         read_context.shrink_input_buffer()
         return obj
 
-    def _read_missing_field_value(self, read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref):
+    def _read_missing_field_value(
+        self,
+        read_context,
+        serializer,
+        is_nullable,
+        is_dynamic,
+        is_basic,
+        is_tracking_ref,
+        is_compatible_scalar,
+    ):
         previous = self.type_resolver._allow_unregistered_typedef
         self.type_resolver._allow_unregistered_typedef = True
         try:
-            return self._read_field_value(read_context, serializer, is_nullable, is_dynamic, is_basic, is_tracking_ref)
+            return self._read_field_value(
+                read_context,
+                serializer,
+                is_nullable,
+                is_dynamic,
+                is_basic,
+                is_tracking_ref,
+                is_compatible_scalar,
+            )
         finally:
             self.type_resolver._allow_unregistered_typedef = previous
